third_party/WebKit/LayoutTests/editing/execCommand/insert-image-changing-visibility-crash.html

<!DOCTYPE html>
<html>
<head>
<title>Issue 348283</title>
<script src="../../resources/js-test.js"></script>
</head>
<body>
<script>
description('Should not crash if we load a test case from crbug.com/348283.');

window.jsTestIsAsync = true;

window.addEventListener('message', didReceiveMessage, false);

var iframe = document.createElement('iframe');
iframe.src = 'resources/insert-image-changing-visibility-crash-iframe.html';
document.body.appendChild(iframe);

function didReceiveMessage(event)
{
    shouldBeEqualToString('event.data', 'FINISH');
    document.body.removeChild(iframe);
    testPassed('Did not crash.');
    window.finishJSTest();
}
</script>
</body>
</html>
